Transaction_Monitoring/requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
flask==2.3.3
flask-sse==0.2.1